requirements.txt
Django==1.8.1
djangorestframework==3.1.1
model-mommy==1.2.4
psycopg2==2.6
selenium==2.45.0
